Add rest API

+++ cart/src/main/java/com/example/cart/cart/CartRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 @Repository
 public interface CartRepository extends JpaRepository<Cart, Integer> {
-    List<Cart> findByUserId(Long userId);
+    List<Cart> findByUserId(Integer userId);
 
 }
